requirements.txt
streamlit==1.50.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
pillow==11.3.0
